+++ projeto/projeto-final/src/main/java/com/tech4me/projetofinal/service/doce/DoceService.java
@@ -7,9 +7,9 @@ import com.tech4me.projetofinal.model.doce.Doce;
 
 public interface DoceService {
     List<Doce> obterTodos();
-    Optional<Doce> obterPorId(Integer id);
-    Doce atualizar(Integer id, Doce Doce);
+    Optional<Doce> obterPorId(Long id);
+    Doce atualizar(Long id, Doce Doce);
     Doce cadastrar(Doce Doce);
-    Doce deletar(Integer id);
+    Doce deletar(Long id);
    
 }
